Make Tiles data request default to "mvt" format
QGIS does not set "accept" headers which result in a non-None content
type being negotiated for the response. However, the MVT view does not
seem to pay any heed to the `format`, other than passing it as a file
extension to pg_tileserv, so defaulting to "mvt" for the format triggers
its logic to just use the extension configured for the Provider.