Update file extension requesting criterion for MVT-based Tiles implementation
The test did not seem to make sense, since a pg_tileserv URL will always
have a `.` in the path (it is used as a separator between schema and
table name).

This may not be the correct solution, but in case of a URL, more complex
parsing would be necessary to correctly determine whether a file
extension was specified, if this is indeed what the test should be
looking at.